🚧 moving non-view methods from helper to concern
